fix(compare): address review on compare PR (#1206)
- Guard the BCF-from-change submit against double-submit (in-flight ref) and
  wrap the non-idempotent topic create + viewpoint capture in try/catch/finally
  so a failure can't leave an unhandled rejection or a stuck form.
- Neutralise CSV formula injection in the change-report export: values led by
  = + - @ / tab / CR are prefixed with an apostrophe so spreadsheets treat
  attacker-influenced model/element names as text, not formulas.
- Split ComparePanel (was 621 lines) under the ~400-line house rule: extract
  the change-detail view, BCF-from-change affordance, results list + count
  badges, and the shared row helpers into apps/viewer/.../compare/*.

Tests: CSV formula-escaping case added; compare suite + viewer build green.
