+++ apps/viewer/src/components/viewer/ComparePanel.tsx
@@ -9,10 +9,9 @@
  * `useCompareOverlay`) and listed here. Row click selects + frames the element.
  */
 
-import { useEffect, useMemo, useState } from 'react';
-import { GitCompareArrows, Plus, Minus, PencilLine, Loader2, Play, X, Trash2, Download, MessageSquarePlus, CheckCircle2 } from 'lucide-react';
+import { useEffect, useMemo, useRef, useState } from 'react';
+import { GitCompareArrows, Loader2, Play, X, Trash2, Download } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import { ScrollArea } from '@/components/ui/scroll-area';
 import { cn } from '@/lib/utils';
 import { useViewerStore } from '@/store';
 import { useCompare } from '@/hooks/useCompare';
@@ -20,11 +19,14 @@ import { useCompareOverlay } from '@/hooks/useCompareOverlay';
 import { useBCF } from '@/hooks/useBCF';
 import { posthog } from '@/lib/analytics';
 import { createBCFProject, createBCFTopic, type BCFTopic } from '@ifc-lite/bcf';
-import { COMPARE_COLORS, type RGBA } from '@/lib/compare/overlay';
+import { COMPARE_COLORS } from '@/lib/compare/overlay';
 import type { CompareRef } from '@/lib/compare/buildFingerprints';
-import { describeChange, type ChangeDetail, type FieldDelta, type GeometrySummary } from '@/lib/compare/describeChange';
+import { describeChange, type ChangeDetail } from '@/lib/compare/describeChange';
 import { downloadCompareReport } from '@/lib/compare/exportReport';
-import { BCFCreateTopicForm } from './bcf/BCFCreateTopicForm';
+import { ChangeDetailView } from './compare/ChangeDetailView';
+import { BcfFromChange } from './compare/BcfFromChange';
+import { CompareResultsList, CountBadge, LISTED_STATES, type CompareBucket } from './compare/CompareResultsList';
+import type { CompareRow } from './compare/changeRow';
 import type { DiffScope, DiffState, DiffEntry } from '@ifc-lite/diff';
 
 interface ComparePanelProps {
@@ -37,65 +39,12 @@ const SCOPES: { id: DiffScope; label: string }[] = [
   { id: 'geometry', label: 'Geometry' },
 ];
 
-/** States listed in the panel (unchanged only affects 3D ghosting). */
-const LISTED_STATES: { state: Exclude<DiffState, 'unchanged'>; label: string; color: RGBA; Icon: typeof Plus }[] = [
-  { state: 'modified', label: 'Changed', color: COMPARE_COLORS.modified, Icon: PencilLine },
-  { state: 'added', label: 'Added', color: COMPARE_COLORS.added, Icon: Plus },
-  { state: 'deleted', label: 'Deleted', color: COMPARE_COLORS.deleted, Icon: Minus },
-];
-
 /** Cap rows rendered per group so a huge diff can't stall the DOM. */
 const MAX_ROWS_PER_GROUP = 1000;
-
-function rgbaCss([r, g, b, a]: RGBA): string {
-  return `rgba(${Math.round(r * 255)}, ${Math.round(g * 255)}, ${Math.round(b * 255)}, ${a})`;
-}
-
-interface CompareRow {
-  key: string;
-  ifcType: string;
-  name: string;
-  state: DiffState;
-  changeKinds: string[];
-  ref: CompareRef;
-}
 
 /** The side actually drawn for an entry: base for deletions, head otherwise. */
 function renderRef(entry: DiffEntry<CompareRef>): CompareRef | undefined {
   return (entry.state === 'deleted' ? entry.base?.ref : entry.head?.ref) ?? entry.base?.ref;
-}
-
-/** A short human change label for a row (added / deleted / the change kinds). */
-function changeLabel(row: CompareRow): string {
-  if (row.state === 'added') return 'added';
-  if (row.state === 'deleted') return 'deleted';
-  return row.changeKinds.length ? row.changeKinds.join(' + ') : 'changed';
-}
-
-/** Pre-fill a BCF topic title + description from a detected change (#1199). */
-function bcfTextFromChange(row: CompareRow, detail: ChangeDetail | null): { title: string; description: string } {
-  const typeLabel = row.ifcType.replace(/^Ifc/, '');
-  const name = row.name || typeLabel;
-  const title = `${typeLabel} "${name}" - ${changeLabel(row)}`;
-  const lines: string[] = [
-    `Detected in model comparison: ${changeLabel(row)}.`,
-    row.key.startsWith('missing:') ? '' : `GlobalId: ${row.key}`,
-  ];
-  if (detail?.geometry) {
-    if (detail.geometry.movedDistance > 0) lines.push(`Moved ${detail.geometry.movedDistance.toFixed(3)} m.`);
-    if (detail.geometry.reshaped) lines.push('Bounding box reshaped.');
-  }
-  if (detail?.data?.length) {
-    lines.push('', 'Data changes:');
-    for (const d of detail.data.slice(0, 20)) {
-      const where = d.group ? `${d.group} / ${d.name}` : d.name;
-      if (d.kind === 'changed') lines.push(`- ${where}: ${d.before ?? '-'} -> ${d.after ?? '-'}`);
-      else if (d.kind === 'added') lines.push(`- ${where}: added ${d.after ?? ''}`.trimEnd());
-      else lines.push(`- ${where}: removed`);
-    }
-    if (detail.data.length > 20) lines.push(`- ... and ${detail.data.length - 20} more`);
-  }
-  return { title, description: lines.filter((l, i) => l !== '' || i > 0).join('\n') };
 }
 
 export function ComparePanel({ onClose }: ComparePanelProps) {
@@ -119,6 +68,8 @@ export function ComparePanel({ onClose }: ComparePanelProps) {
 
   const [bcfFormOpen, setBcfFormOpen] = useState(false);
   const [bcfCreatedTitle, setBcfCreatedTitle] = useState<string | null>(null);
+  // Guards the non-idempotent BCF create against a double-submit (#1208 review).
+  const bcfSubmitInFlight = useRef(false);
 
   const modelList = useMemo(() => Array.from(models.values()), [models]);
 
@@ -148,9 +99,9 @@ export function ComparePanel({ onClose }: ComparePanelProps) {
   // the per-model store (the engine result carries only type + key), so we
   // look them up here via each entry's ref.
   const groups = useMemo(() => {
-    const empty = new Map<DiffState, { rows: CompareRow[]; truncated: number }>();
+    const empty = new Map<DiffState, CompareBucket>();
     if (!result) return empty;
-    const out = new Map<DiffState, { rows: CompareRow[]; truncated: number }>();
+    const out = new Map<DiffState, CompareBucket>();
     for (const { state } of LISTED_STATES) out.set(state, { rows: [], truncated: 0 });
 
     for (const entry of result.diff.entries) {
@@ -221,34 +172,42 @@ export function ComparePanel({ onClose }: ComparePanelProps) {
   // currently-framed element, and stay in the compare view. Priority / type are
   // set in the form; status / author / GlobalId come along automatically.
   const submitBcfFromChange = async (data: Partial<BCFTopic>) => {
-    const state = useViewerStore.getState();
-    if (!state.bcfProject) {
-      const first = modelList[0]?.name?.replace(/\.(ifc|ifczip)$/i, '') || 'Comparison';
-      state.setBcfProject(createBCFProject({ name: `${first}_Issues` }));
+    if (bcfSubmitInFlight.current) return;
+    bcfSubmitInFlight.current = true;
+    try {
+      const state = useViewerStore.getState();
+      if (!state.bcfProject) {
+        const first = modelList[0]?.name?.replace(/\.(ifc|ifczip)$/i, '') || 'Comparison';
+        state.setBcfProject(createBCFProject({ name: `${first}_Issues` }));
+      }
+      const topic = createBCFTopic({
+        title: data.title || 'Untitled',
+        description: data.description,
+        author: state.bcfAuthor,
+        topicType: data.topicType,
+        topicStatus: 'Open',
+        priority: data.priority,
+      });
+      useViewerStore.getState().addTopic(topic);
+      const viewpoint = await createViewpointFromState({
+        includeSnapshot: true,
+        includeSelection: true,
+        includeHidden: false,
+      });
+      if (viewpoint) useViewerStore.getState().addViewpoint(topic.guid, viewpoint);
+      posthog.capture('bcf_topic_created', {
+        source: 'compare',
+        topic_type: topic.topicType,
+        priority: topic.priority,
+        has_viewpoint: Boolean(viewpoint),
+      });
+      setBcfFormOpen(false);
+      setBcfCreatedTitle(topic.title);
+    } catch (error) {
+      console.error('[compare] failed to create BCF issue from change', error);
+    } finally {
+      bcfSubmitInFlight.current = false;
     }
-    const topic = createBCFTopic({
-      title: data.title || 'Untitled',
-      description: data.description,
-      author: state.bcfAuthor,
-      topicType: data.topicType,
-      topicStatus: 'Open',
-      priority: data.priority,
-    });
-    useViewerStore.getState().addTopic(topic);
-    const viewpoint = await createViewpointFromState({
-      includeSnapshot: true,
-      includeSelection: true,
-      includeHidden: false,
-    });
-    if (viewpoint) useViewerStore.getState().addViewpoint(topic.guid, viewpoint);
-    posthog.capture('bcf_topic_created', {
-      source: 'compare',
-      topic_type: topic.topicType,
-      priority: topic.priority,
-      has_viewpoint: Boolean(viewpoint),
-    });
-    setBcfFormOpen(false);
-    setBcfCreatedTitle(topic.title);
   };
 
   return (
@@ -375,64 +334,13 @@ export function ComparePanel({ onClose }: ComparePanelProps) {
           )}
 
           {/* Results list */}
-          <ScrollArea className="flex-1 min-h-0">
-            {!result ? (
-              <div className="p-4 text-sm text-muted-foreground">
-                Run a comparison to see added, changed, and deleted elements.
-              </div>
-            ) : (
-              <div className="p-2 space-y-3">
-                {LISTED_STATES.map(({ state, label, color, Icon }) => {
-                  const bucket = groups.get(state);
-                  if (!bucket || bucket.rows.length === 0) return null;
-                  return (
-                    <div key={state}>
-                      <div className="flex items-center gap-1.5 px-1 py-1 text-xs font-medium">
-                        <Icon className="h-3.5 w-3.5" style={{ color: rgbaCss(color) }} />
-                        <span>{label}</span>
-                        <span className="text-muted-foreground">({bucket.rows.length + bucket.truncated})</span>
-                      </div>
-                      <div className="space-y-0.5">
-                        {bucket.rows.map((row) => (
-                          <button
-                            key={row.key}
-                            onClick={() => focusEntry(row)}
-                            className={cn(
-                              'w-full text-left rounded px-2 py-1 flex items-center gap-2 hover:bg-muted transition-colors min-w-0',
-                              selectedKey === row.key && 'bg-muted',
-                            )}
-                          >
-                            <span
-                              className="h-2.5 w-2.5 rounded-sm shrink-0"
-                              style={{ backgroundColor: rgbaCss(color) }}
-                            />
-                            <span className="min-w-0 flex-1 truncate text-xs">
-                              {row.name || row.ifcType}
-                            </span>
-                            <span className="shrink-0 text-[10px] text-muted-foreground">
-                              {state === 'modified' && row.changeKinds.length > 0
-                                ? row.changeKinds.join(' · ')
-                                : row.ifcType.replace(/^Ifc/, '')}
-                            </span>
-                          </button>
-                        ))}
-                        {bucket.truncated > 0 && (
-                          <p className="px-2 py-1 text-[10px] text-muted-foreground">
-                            +{bucket.truncated} more not shown
-                          </p>
-                        )}
-                      </div>
-                    </div>
-                  );
-                })}
-                {counts && counts.added + counts.modified + counts.deleted === 0 && (
-                  <div className="p-3 text-sm text-muted-foreground">
-                    No differences in scope “{result.scope}”. The models match.
-                  </div>
-                )}
-              </div>
-            )}
-          </ScrollArea>
+          <CompareResultsList
+            result={result}
+            groups={groups}
+            counts={counts}
+            selectedKey={selectedKey}
+            onFocus={focusEntry}
+          />
 
           {/* What-changed detail for the selected element */}
           {detail && selectedRow && <ChangeDetailView row={selectedRow} detail={detail} />}
@@ -453,169 +361,6 @@ export function ComparePanel({ onClose }: ComparePanelProps) {
           )}
         </>
       )}
-    </div>
-  );
-}
-
-/** "Create BCF issue" affordance shown under the focused change (#1199). */
-function BcfFromChange({
-  row,
-  detail,
-  author,
-  open,
-  createdTitle,
-  onStart,
-  onCancel,
-  onSubmit,
-  onOpenBcfPanel,
-}: {
-  row: CompareRow;
-  detail: ChangeDetail | null;
-  author: string;
-  open: boolean;
-  createdTitle: string | null;
-  onStart: () => void;
-  onCancel: () => void;
-  onSubmit: (topic: Partial<BCFTopic>) => void;
-  onOpenBcfPanel: () => void;
-}) {
-  const prefill = useMemo(() => bcfTextFromChange(row, detail), [row, detail]);
-
-  if (createdTitle) {
-    return (
-      <div className="border-t border-border shrink-0 px-3 py-2.5 flex items-center gap-2 text-xs">
-        <CheckCircle2 className="h-4 w-4 text-[#9ece6a] shrink-0" />
-        <span className="min-w-0 truncate">BCF issue created: “{createdTitle}”</span>
-        <Button variant="outline" size="sm" className="ml-auto h-7 px-2 text-xs shrink-0" onClick={onOpenBcfPanel}>
-          Open BCF
-        </Button>
-      </div>
-    );
-  }
-
-  if (open) {
-    return (
-      <div className="border-t border-border shrink-0 max-h-[60%] overflow-auto">
-        <BCFCreateTopicForm
-          author={author}
-          initialTitle={prefill.title}
-          initialDescription={prefill.description}
-          onSubmit={onSubmit}
-          onCancel={onCancel}
-        />
-      </div>
-    );
-  }
-
-  return (
-    <div className="border-t border-border shrink-0 px-3 py-2.5">
-      <Button variant="outline" size="sm" className="w-full gap-1.5 text-xs" onClick={onStart}>
-        <MessageSquarePlus className="h-3.5 w-3.5" />
-        Create BCF issue
-      </Button>
-    </div>
-  );
-}
-
-/** Per-element "what changed" — geometry move/reshape + data field deltas. */
-function ChangeDetailView({ row, detail }: { row: CompareRow; detail: ChangeDetail }) {
-  return (
-    <div className="border-t border-border shrink-0 max-h-[42%] overflow-auto">
-      <div className="px-3 pt-2.5 pb-1.5 flex items-center gap-1.5 sticky top-0 bg-background">
-        <PencilLine className="h-3.5 w-3.5 text-primary shrink-0" />
-        <span className="text-xs font-semibold truncate">{row.name || row.ifcType}</span>
-        <span className="ml-auto text-[10px] text-muted-foreground shrink-0">{row.ifcType.replace(/^Ifc/, '')}</span>
-      </div>
-      <div className="px-3 pb-3 space-y-2.5 text-xs">
-        {detail.geometry && (
-          <div className="space-y-1">
-            <div className="text-[11px] font-medium text-muted-foreground uppercase tracking-wide">Geometry</div>
-            <GeometryDetail summary={detail.geometry} />
-          </div>
-        )}
-        {detail.data.length > 0 ? (
-          <div className="space-y-1">
-            <div className="text-[11px] font-medium text-muted-foreground uppercase tracking-wide">
-              Data <span className="text-muted-foreground/70">({detail.data.length})</span>
-            </div>
-            <div className="space-y-1">
-              {detail.data.map((d, i) => <FieldDeltaRow key={i} delta={d} />)}
-            </div>
-          </div>
-        ) : detail.dataOnlyGeometric ? (
-          <div className="text-[11px] text-muted-foreground italic">
-            Data fingerprint differs but no field-level change could be pinpointed.
-          </div>
-        ) : !detail.geometry ? (
-          <div className="text-[11px] text-muted-foreground italic">No field-level detail available.</div>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-function GeometryDetail({ summary }: { summary: GeometrySummary }) {
-  const moved = summary.movedDistance > 0;
-  const fmt = (n: number) => (Math.abs(n) < 1e-3 ? '0' : n.toFixed(Math.abs(n) >= 1 ? 2 : 3));
-  const signed = (n: number) => (n > 0 ? `+${fmt(n)}` : fmt(n));
-  const headline = summary.reshaped ? (moved ? 'Reshaped + moved' : 'Reshaped') : moved ? 'Moved' : 'Geometry changed';
-  return (
-    <div className="rounded border border-border/60 px-2 py-1.5 space-y-0.5">
-      <div className="font-medium">{headline}</div>
-      {moved && (
-        <div className="text-muted-foreground tabular-nums">
-          {fmt(summary.movedDistance)} m
-          <span className="text-muted-foreground/70">
-            {' '}(Δx {fmt(summary.delta.x)}, Δy {fmt(summary.delta.y)}, Δz {fmt(summary.delta.z)})
-          </span>
-        </div>
-      )}
-      {summary.reshaped && (
-        <div className="text-muted-foreground tabular-nums">
-          size{' '}
-          <span className="text-muted-foreground/70">
-            (Δx {signed(summary.sizeDelta.x)}, Δy {signed(summary.sizeDelta.y)}, Δz {signed(summary.sizeDelta.z)}) m
-          </span>
-        </div>
-      )}
-      {!moved && !summary.reshaped && (
-        <div className="text-muted-foreground/70 text-[11px]">
-          Shape hash differs but the element’s position and size are unchanged.
-        </div>
-      )}
-    </div>
-  );
-}
-
-function FieldDeltaRow({ delta }: { delta: FieldDelta }) {
-  const kindColor: Record<FieldDelta['kind'], string> = {
-    changed: 'text-[#e0af68]',
-    added: 'text-[#9ece6a]',
-    removed: 'text-[#f7768e]',
-  };
-  return (
-    <div className="rounded border border-border/40 px-2 py-1">
-      <div className="flex items-baseline gap-1.5 min-w-0">
-        {delta.group && <span className="text-[10px] text-muted-foreground shrink-0 truncate max-w-[40%]">{delta.group}</span>}
-        <span className="text-[11px] font-medium truncate">{delta.name}</span>
-        <span className={cn('ml-auto text-[10px] shrink-0', kindColor[delta.kind])}>{delta.kind}</span>
-      </div>
-      <div className="flex items-center gap-1.5 text-[11px] tabular-nums mt-0.5 min-w-0">
-        <span className="text-muted-foreground line-through truncate max-w-[45%]">{delta.before ?? '—'}</span>
-        <span className="text-muted-foreground/60 shrink-0">→</span>
-        <span className="truncate max-w-[45%]">{delta.after ?? '—'}</span>
-      </div>
-    </div>
-  );
-}
-
-function CountBadge({ label, value, color }: { label: string; value: number; color: RGBA }) {
-  return (
-    <div className="flex flex-col items-center gap-0.5">
-      <span className="text-sm font-semibold tabular-nums" style={{ color: rgbaCss([color[0], color[1], color[2], 1]) }}>
-        {value.toLocaleString()}
-      </span>
-      <span className="text-[10px] text-muted-foreground">{label}</span>
     </div>
   );
 }
